requirements.txt
# Core Django Framework (Python 3.9 compatible)
Django==4.2.16
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.0

# Google Earth Engine
earthengine-api==0.1.386

# Google Authentication & OAuth
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-api-core==2.25.1
google-api-python-client==2.177.0

# HTTP Requests
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.1

# Database (PostgreSQL - optional)
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0

# JWT
PyJWT==2.10.1

# ASGI (for async support)
asgiref==3.9.1

# Cloud Storage (for static files on GCS)
django-storages[google]==1.14.2

# AI & Data Processing
google-cloud-aiplatform==1.38.1
pandas==2.1.4
